Moved the dataset class into the data_ingestion/infrastructure new folder hierarchy.
