+++ include/lbann/data_coordinator/data_coordinator.hpp
@@ -27,9 +27,9 @@
 #ifndef LBANN_DATA_COORDINATOR_HPP
 #define LBANN_DATA_COORDINATOR_HPP
 
+#include "lbann/data_ingestion/infrastructure/dataset.hpp"
 #include "lbann/data_ingestion/readers/metadata.hpp"
 #include "lbann/data_ingestion/readers/utils/input_data_type.hpp"
-#include "lbann/utils/dataset.hpp"
 #include "lbann/utils/threads/thread_pool.hpp"
 
 // #ifdef LBANN_HAS_DISTCONV
