+++ include/lbann/data_ingestion/infrastructure/dataset.hpp
