+++ src/data_ingestion/infrastructure/CMakeLists.txt
@@ -23,10 +23,11 @@
 ## implied. See the License for the specific language governing
 ## permissions and limitations under the license.
 ################################################################################
-
-# Add the subdirectories
-add_subdirectory(infrastructure)
-add_subdirectory(readers)
+# Add the source files for this directory
+set_full_path(THIS_DIR_SOURCES
+  dataset.cpp
+)
 
 # Propagate the files up the tree
 set(SOURCES "${SOURCES}" "${THIS_DIR_SOURCES}" PARENT_SCOPE)
+set(GPU_SOURCES "${GPU_SOURCES}" "${THIS_DIR_CU_SOURCES}" PARENT_SCOPE)
