+++ src/utils/CMakeLists.txt
@@ -29,7 +29,6 @@ set_full_path(THIS_DIR_SOURCES
   argument_parser.cpp
   commify.cpp
   cudnn.cpp
-  dataset.cpp
   description.cpp
   environment_variable.cpp
   exception.cpp
